fix wrong vec type

+++ Engine/src/Window.cpp
@@ -4,11 +4,11 @@
 
 namespace basilisk
 {
-    Window::Window(const std::string& windowName, const glm::vec2 size) :
+    Window::Window(const std::string& windowName, const glm::ivec2 size) :
         Size(size)
     {
         /* Create a windowed mode window and its OpenGL context */
-        this->WindowP = glfwCreateWindow(static_cast<int>(size.x), static_cast<int>(size.y), windowName.c_str(), nullptr, nullptr);
+        this->WindowP = glfwCreateWindow(size.x, size.y, windowName.c_str(), nullptr, nullptr);
         
         if (!this->WindowP)
         {
@@ -16,6 +16,7 @@ namespace basilisk
             throw FailedWindowCreation();
         }
         glfwMakeContextCurrent(this->WindowP);
+        
 
     }
